+++ .env.template
@@ -4,7 +4,7 @@ MOD_IS_RELEASE=false
 PUB_DRY_RUN=true
 PUB_MAVEN_ENABLE=false
 PUB_MODS_ENABLE=false
-PUB_GITHUB_RELEASES=false
+PUB_GITHUB_ENABLE=false
 
 # Platform Tokens: replace these with your actual tokens before use
 # Modrinth PAT: https://modrinth.com/settings/pats
